% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Data Preparation}
\usage{
data_prep(df, xgb = F)
}
\arguments{
\item{df}{Input data frame object for processing}
}
\description{
Convert CrimeBB data frame to df for model training
Assumes the following columns: post
Assumes the following columns at positions 9,10,11,12,13,17,24,26,27,31,32:
hasImage, hasLink, hasCode, hasIframe, hasCitation, addressOP, forumID, firstPost, threadCount, authorIsOP, citesOP
Dependencies: tm, xgboost
}
\examples{
data_prep(my.df)
}
\keyword{learning}
\keyword{machine}
\keyword{preprocessing,}
