% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_processing.R
\name{initial_processing}
\alias{initial_processing}
\title{Initial Processing}
\usage{
initial_processing(df)
}
\arguments{
\item{df}{Input data frame object for processing}
}
\description{
First step processing of data frames from the CrimeBB database
Assumes at least the following columns: author, threadOP, postID, threadID, post, threadTitle
Depends on stringr library for str_match_all()
}
\examples{
initial_processing(my.df)
}
\keyword{preprocessing}
