% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_align.R
\name{matrix_align}
\alias{matrix_align}
\title{Align Document-Term Matrices}
\usage{
matrix_align(mtx1, mtx2)
}
\arguments{
\item{df}{Two matrix objects, where the first matrix has at least as many columns as matrix 2}
}
\description{
Align two document-term matrices where 1st dtm was used to train stat.model and 2nd dtm is new test data
Returns matrix with columns ordered alphabetically by column header
}
\examples{
matrix_align(mtx1, mtx2)
}
\keyword{learning}
\keyword{machine}
\keyword{preprocessing,}
