% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_addressee.R
\name{predict_addressee}
\alias{predict_addressee}
\title{Predict Addressee}
\usage{
predict_addressee(df, model = NULL, train.dtm = NULL, labs = NULL)
}
\arguments{
\item{df}{Input data frame object for processing}
}
\description{
Identifies most likely addressee on basis of position in thread, isThreadOP or not, cited post (if present)
Assumes at least the following columns: postID, post, author, threadID, threadOP; firstPost, postNumber, hasCitation, citesPostID, citesUser (last 5 columns come from initial_processing() function).
}
\examples{
predict_addressee(my.df)
}
\keyword{addressee,}
\keyword{analysis}
\keyword{conversation}
