% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_intent.R
\name{predict_intent}
\alias{predict_intent}
\title{Predict Intent}
\usage{
predict_intent(df, model = NULL, train.dtm = NULL, labs = NULL)
}
\arguments{
\item{df}{Input data frame object for processing, plus optional trained model, document-term matrix and list of labels used in model training.}
}
\description{
Predict the author's intent in writing a CrimeBB post.
Assumes at least the following columns: post, firstPost, sentiment (last 2 columns come from initial_processing() function).
Dependencies: tm, LiblineaR
}
\examples{
predict_intent(my.df)
}
\keyword{author}
\keyword{intent,}
\keyword{sentiment}
