% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_posttype.R
\name{predict_posttype}
\alias{predict_posttype}
\title{Predict Post Type}
\usage{
predict_posttype(df, model = NULL, train.dtm = NULL, labs = NULL)
}
\arguments{
\item{df}{Input data frame object for processing, and optional trained model, document-term matrix and list of labels used in model training.}
}
\description{
Predict the type of each post in a CrimeBB dataset.
Assumes at least the following columns: post, bboardTitle, firstPost (last column comes from initial_processing() function).
Dependencies: tm, LiblineaR
}
\examples{
predict_posttype(my.df)
}
\keyword{post}
\keyword{type}
