% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repvote_matching.R
\name{repvote_matching}
\alias{repvote_matching}
\title{Reputation Vote Matching}
\usage{
repvote_matching(df, memb.df, rep.df, execute = T)
}
\arguments{
\item{df}{Input data frame object for processing}
}
\description{
Match user posts with reputation votes.
Assumes at least the following columns: timestamp, authorID, addressee
And a data frame for members and reputation votes already downloaded from CrimeBB database.
Depends on tidytext library
}
\examples{
repvote_matching(my.df)
}
\keyword{reputation}
\keyword{votes}
