% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_stats.R
\name{user_stats}
\alias{user_stats}
\title{User Stats}
\usage{
user_stats(df.in, thread.df, memb.df, rep.df)
}
\arguments{
\item{df}{Input data frame object for processing}
}
\description{
Summary user stats based on posts extracted from CrimeBB database.
Assumes a posts data frame with at least the following columns: repVoteReceived, timestamp, threadID, firstPost, bboardID, tokenCount, sentiment, hasImage, hasCode, hasLink, hasIframe, hasCitation (all pre-processed)
And a subset of the member table from CrimeBB.
And reputation vote data frame downloaded from CrimeBB database.
}
\examples{
user_stats(my.df)
}
\keyword{analysis}
\keyword{user}
