% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Data Preparation}
\usage{
data_prep(df, xgb = F)
}
\arguments{
\item{df}{Input data frame object for processing}
}
\description{
Convert CrimeBB data frame to df for model training
Assumes the following columns: post
Assumes the following columns at positions 10, 14, 15, 17:21
bboardID, postNumber, firstPost, hasImage, hasCode, hasLink, hasIframe, hasCitation
Dependencies: tm, xgboost
}
\examples{
data_prep(my.df)
}
\keyword{learning}
\keyword{machine}
\keyword{preprocessing,}
