% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_align.R
\name{matrix_align}
\alias{matrix_align}
\title{Align Document-Term Matrices}
\usage{
matrix_align(df1, df2)
}
\arguments{
\item{df}{Two data frames}
}
\description{
Align two data frames where 1st df was used to train stat.model and 2nd df is target data
Returns matrix with columns ordered alphabetically by column header
Assumes the following columns: post
}
\examples{
matrix_align(my.df1, my.df2)
}
\keyword{learning}
\keyword{machine}
\keyword{preprocessing,}
